Fix compiler warnings about implicit conversions from 64 to 32 bits, and missing braces in initializers.

+++ Development/Source/Platform/Cocoa/E3CocoaStackCrawl.c
@@ -424,7 +424,7 @@ const MachOHeader *MachOMapArch(const MachOHeader *mh,const char *path,size_t *s
 				(mh32.cpusubtype == mh->cpusubtype))
 			{
 				offset = 0;
-				*size = finfo.st_size;
+				*size = (size_t)finfo.st_size;
 			}
 		}
